+++ application/src/main/java/fr/delivrooom/application/tsp/example/GrapheComplet.java
@@ -1,4 +1,4 @@
-package fr.delivrooom.application.tsp;
+package fr.delivrooom.application.tsp.example;
 
 import fr.delivrooom.application.model.Graphe;
